+++ src/entity/beacon.rs
@@ -1,6 +1,6 @@
 use crate::traits::entity;
 use crate::traits::draw::{GenericShape, ShapeVariant};
-use crate::levels::MapIdx;
+use crate::levels::map::MapIdx;
 use crate::misc::point2d::Point2;
 use crate::game::consts::{
     BEACON_SIZE,
