+++ src/levels/mod.rs
@@ -1,2 +1,3 @@
-pub use self::level::{Level, MapIdx};
+pub use self::level::Level;
 mod level;
+pub mod map;
